backend/requirements.txt
# Pixova AI - Production Dependencies (Minimal)
# Install: pip install -r requirements.txt

# ===========================================
# CORE FRAMEWORK
# ===========================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# ===========================================
# API CLIENT
# ===========================================
openai>=1.12.0
httpx>=0.26.0

# ===========================================
# CONFIGURATION
# ===========================================
python-dotenv>=1.0.0

# ===========================================
# IMAGE PROCESSING (lightweight)
# ===========================================
Pillow>=10.0.0
requests>=2.31.0